fix(agents): preserve terminal-only proxy tool calls
